Tag availability query tests
